+++ CHANGELOG.md
@@ -46,6 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(decode): Added `UserHeader` to `GMSOLAccountData`.
 - sdk(sdk): Added support for `update_closed_state` instruction.
 - sdk(sdk): Added support for `set_market_config_updatable` instruction.
+- sdk(sdk): Added support for `initialize` instruction for liquidity provider program.
+- cli: Added `lp init-lp` command. 
 
 ### Changed